# swagger.yaml
openapi: 3.0.0
info:
  title: Chat Api
  contact: {}
  version: "1.0"
servers:
  - url: https://chat-api-06.herokuapp.com/api/v1/
    variables: {}
paths:
...
  /messages/{id}:
    get:
      tags:
        - Messages
      summary: Get conversation messages
      operationId: Getconversationmessages
      parameters:
        - name: x-auth-token
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    delete:
      tags:
        - Messages
      summary: Delete message
      operationId: Deletemessage
      parameters:
        - name: x-auth-token
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /messages:
    post:
      tags:
        - Messages
      summary: Send new message
      operationId: Sendnewmessage
      parameters:
        - name: x-auth-token
          in: header
          description: ""
          required: true
          style: simple
          schema:
            type: string
            example:
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
                - conversationId
              type: object
              properties:
                conversationId:
                  type: string
                  example: 628abdc4fed7e82d54ca8c04
                text:
                  type: string
                  example: "hi "
                image:
                  type: string
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
components:
...
  securitySchemes:
    auth-token:
      type: apiKey
      name: x-auth-token
      in: header
security:
  - auth-token: []
...